requirements.txt
cffi==1.15.1
cryptography==38.0.4
greenlet==2.0.1
pycparser==2.21
PyYAML==6.0
SQLAlchemy==1.4.45
